backend/decision_tree_layer/requirements.txt
kafka-python==2.1.5
psycopg2-binary==2.9.10
requests==2.32.3